Rename ProfileLink1 folder to ProfileLink
